pyproject.toml
[project]
name = "kafka-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kafka-python-ng>=2.2.3",
]
